Add Yogurt ($YOG) to the token list
Yogurt is a HRC-20 meme token on Harmony. But with a value, which gets defined by memes of yogurt and other funny stuff about crypto.

More about YOG: https://www.yogurt.rocks

It would be a huge step for the non-animal oriented meme token, to get listed on the official community token list of LootSwap. Thanks!